Making Context usable client-side.
Previously, the Context structure was used exclusively by C++ since it
contained a reference to a C++ object, but this prevented C clients from
accessing its error reporting fields.

Changed the C++ object reference to a void* and extern'd it for C
clients to use. Ramdis now must convert the void* pointer to a
RAMCloud::RamCloud* pointer before use.