Added RPUSH test.
Currently this test will not necessarily pass because delete does not
properly delete lists. Therefore whichever PUSH test runs first will
work fine and then the next will not, because they are operating on the
same key and the first one to run did not properly delete the list.

Fix coming soon.