Changed object key encodings.
Previously keys were encoded in such a way that it was possible for a
string objects and list objects to exists separately under the same key.
So if the user invoked string commands on the key, it would operate on
the stored string for that key. If the user invoked list commands on the
key, it would operate on the stored list for that key. This does not
match Redis behavior, however. Keys map to a single, typed object in
Redis.

The key encodings have now been changed so that list ops on a string or
vice versa will throw an error.