Added object metadata to internal storage format.
The previous commit made key encodings refer to unique objects. This
commit makes it so that operations can detect the data type of objects
based on what is read out of RAMCloud. Previously the data type was
inferred from the operation. Now operations like delete can figure out
the data type and, thus, the format and how to properly delete the
object be it a string, list, hash, set, etc.