Fixed bug in transaction retry mechanism.
Previously I applied a simple while loop and continue statement
construct to retry failed transactions, but overlooked that some of
those continue statements were inside loops themselves (major >_<!).
This resulted in the RAMCloud client library complaining that the
program was trying to use an already committed transaction context.

To fix this I restructured the code so that transactions are not
attempted inside loops. I was lucky to be able to do this with the
present code, but there may be situations where avoiding tx.commit() in
loops becomes too awkward or cumbersome. In those cases, I suspect that
using goto statements and a label at the top of the transaction code may
be the simplest approach, although I may risk being eaten by
velociraptors in that case.