Added timeLimit program option.
This parameter sets a time limit on the client threads. This was mainly
added because cluster.py takes a timeout parameter with a default of
30s, and if the clients don't finish before the timeout then the clients
are killed. This ensures that the clients save what they measured before
the timeout. It is also generally useful when the request rate is very
low and we don't want our experiment to take forever.